# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "Shahn ERP foundation", "version": "1.0.0" },
  "components": {
    "securitySchemes": {
      "erpSession": { "type": "apiKey", "in": "cookie", "name": "erp_session" },
      "csrfToken": { "type": "apiKey", "in": "header", "name": "X-CSRF-Token" }
    }
  },
  "paths": {
...
    "/api/v1/access/login": {
      "post": {
        "summary": "Uniform issuer redirect; no company/user existence lookup",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": ["companyCode", "username", "support", "returnPath"],
                "properties": {
                  "companyCode": { "type": "string", "maxLength": 40 },
                  "username": { "type": "string", "maxLength": 80 },
                  "support": { "type": "boolean" },
                  "returnPath": { "type": "string", "maxLength": 200, "pattern": "^/(?!/)" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Issuer redirect URL",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": ["url"],
                  "properties": { "url": { "type": "string", "format": "uri" } }
                }
              }
            }
          },
          "400": {
            "description": "Typed rejection",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": ["code", "messageKey", "commandId", "correlationId"],
                  "properties": {
                    "code": { "type": "string" },
                    "messageKey": { "type": "string" },
                    "commandId": {
                      "anyOf": [{ "type": "string", "format": "uuid" }, { "type": "null" }]
                    },
                    "correlationId": { "type": "string", "format": "uuid" }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Typed rejection",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": ["code", "messageKey", "commandId", "correlationId"],
                  "properties": {
                    "code": { "type": "string" },
                    "messageKey": { "type": "string" },
                    "commandId": {
                      "anyOf": [{ "type": "string", "format": "uuid" }, { "type": "null" }]
                    },
                    "correlationId": { "type": "string", "format": "uuid" }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Typed rejection",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": ["code", "messageKey", "commandId", "correlationId"],
                  "properties": {
                    "code": { "type": "string" },
                    "messageKey": { "type": "string" },
                    "commandId": {
                      "anyOf": [{ "type": "string", "format": "uuid" }, { "type": "null" }]
                    },
                    "correlationId": { "type": "string", "format": "uuid" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/access/session": {
      "get": {
        "summary": "Server-authorized session",
        "security": [{ "erpSession": [] }],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Current authorized response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": ["principalId", "csrfToken", "kind", "companyId"],
                  "properties": {
                    "principalId": { "type": "string", "format": "uuid" },
                    "csrfToken": { "type": "string" },
                    "kind": { "type": "string", "enum": ["staff", "support"] },
                    "companyId": {
                      "anyOf": [{ "type": "string", "format": "uuid" }, { "type": "null" }]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Closed typed validation, lookup, ambiguous command identity or server failure",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": ["code", "messageKey", "commandId", "correlationId"],
                  "properties": {
                    "code": { "type": "string" },
                    "messageKey": { "type": "string" },
                    "commandId": {
                      "anyOf": [{ "type": "string", "format": "uuid" }, { "type": "null" }]
                    },
                    "correlationId": { "type": "string", "format": "uuid" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Reauthentication required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": ["code", "messageKey", "commandId", "correlationId"],
                  "properties": {
                    "code": { "type": "string" },
                    "messageKey": { "type": "string" },
                    "commandId": {
                      "anyOf": [{ "type": "string", "format": "uuid" }, { "type": "null" }]
                    },
                    "correlationId": { "type": "string", "format": "uuid" }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Current authority denies access",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": ["code", "messageKey", "commandId", "correlationId"],
                  "properties": {
                    "code": { "type": "string" },
                    "messageKey": { "type": "string" },
                    "commandId": {
                      "anyOf": [{ "type": "string", "format": "uuid" }, { "type": "null" }]
                    },
                    "correlationId": { "type": "string", "format": "uuid" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Closed typed validation, lookup, ambiguous command identity or server failure",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": ["code", "messageKey", "commandId", "correlationId"],
                  "properties": {
                    "code": { "type": "string" },
                    "messageKey": { "type": "string" },
                    "commandId": {
                      "anyOf": [{ "type": "string", "format": "uuid" }, { "type": "null" }]
                    },
                    "correlationId": { "type": "string", "format": "uuid" }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Closed typed validation, lookup, ambiguous command identity or server failure",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": ["code", "messageKey", "commandId", "correlationId"],
                  "properties": {
                    "code": { "type": "string" },
                    "messageKey": { "type": "string" },
                    "commandId": {
                      "anyOf": [{ "type": "string", "format": "uuid" }, { "type": "null" }]
                    },
                    "correlationId": { "type": "string", "format": "uuid" }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Closed typed validation, lookup, ambiguous command identity or server failure",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": ["code", "messageKey", "commandId", "correlationId"],
                  "properties": {
                    "code": { "type": "string" },
                    "messageKey": { "type": "string" },
                    "commandId": {
                      "anyOf": [{ "type": "string", "format": "uuid" }, { "type": "null" }]
                    },
                    "correlationId": { "type": "string", "format": "uuid" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/access/context": {
      "get": {
        "summary": "Server-authorized context",
        "security": [{ "erpSession": [] }],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Current authorized response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": ["context", "capabilities"],
                  "properties": {
                    "context": {
                      "type": "object",
                      "additionalProperties": false,
                      "required": [
                        "principalId",
                        "principalKind",
                        "sessionId",
                        "companyId",
                        "companyName",
                        "companyActive",
                        "userActive",
                        "issuer",
                        "subject",
                        "displayName",
                        "authorizationRevision",
                        "grants",
                        "assignedBranches",
                        "companyBranches",
                        "supportSessionId",
                        "supportExpiresAt"
                      ],
                      "properties": {
                        "principalId": { "type": "string", "format": "uuid" },
                        "principalKind": { "type": "string", "enum": ["staff", "support"] },
                        "sessionId": { "type": "string", "format": "uuid" },
                        "companyId": { "type": "string", "format": "uuid" },
                        "companyName": { "type": "string" },
                        "companyActive": { "type": "boolean" },
                        "userActive": { "type": "boolean" },
                        "issuer": { "type": "string" },
                        "subject": { "type": "string" },
                        "displayName": { "type": "string" },
                        "authorizationRevision": { "type": "string" },
                        "grants": {
                          "type": "array",
                          "items": {
                            "type": "string",
                            "enum": [
                              "access.users",
                              "access.roles",
                              "intake",
                              "inventory",
                              "expenses",
                              "goods.send",
                              "goods.receive",
                              "treasury.send",
                              "treasury.receive",
                              "tracking",
                              "brand.payout",
                              "employees",
                              "payroll",
                              "reports"
                            ]
                          }
                        },
                        "assignedBranches": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "additionalProperties": false,
                            "required": ["id", "name"],
                            "properties": {
                              "id": { "type": "string", "format": "uuid" },
                              "name": { "type": "string" }
                            }
                          }
                        },
                        "companyBranches": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "additionalProperties": false,
                            "required": ["id", "name"],
                            "properties": {
                              "id": { "type": "string", "format": "uuid" },
                              "name": { "type": "string" }
                            }
                          }
                        },
                        "supportSessionId": {
                          "anyOf": [{ "type": "string", "format": "uuid" }, { "type": "null" }]
                        },
                        "supportExpiresAt": {
                          "anyOf": [{ "type": "string", "format": "date-time" }, { "type": "null" }]
                        }
                      }
                    },
                    "capabilities": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": false,
                        "required": ["id", "title", "route", "policy", "implemented"],
                        "properties": {
                          "id": {
                            "type": "string",
                            "enum": [
                              "access.users",
                              "access.roles",
                              "intake",
                              "inventory",
                              "expenses",
                              "goods.send",
                              "goods.receive",
                              "treasury.send",
                              "treasury.receive",
                              "tracking",
                              "brand.payout",
                              "employees",
                              "payroll",
                              "reports"
                            ]
                          },
                          "title": { "type": "string" },
                          "route": { "type": "string" },
                          "policy": {
                            "type": "string",
                            "enum": ["company", "assigned", "tracking", "treasury", "wallet"]
                          },
                          "implemented": { "type": "boolean" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Closed typed validation, lookup, ambiguous command identity or server failure",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": ["code", "messageKey", "commandId", "correlationId"],
                  "properties": {
                    "code": { "type": "string" },
                    "messageKey": { "type": "string" },
                    "commandId": {
                      "anyOf": [{ "type": "string", "format": "uuid" }, { "type": "null" }]
                    },
                    "correlationId": { "type": "string", "format": "uuid" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Reauthentication required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": ["code", "messageKey", "commandId", "correlationId"],
                  "properties": {
                    "code": { "type": "string" },
                    "messageKey": { "type": "string" },
                    "commandId": {
                      "anyOf": [{ "type": "string", "format": "uuid" }, { "type": "null" }]
                    },
                    "correlationId": { "type": "string", "format": "uuid" }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Current authority denies access",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": ["code", "messageKey", "commandId", "correlationId"],
                  "properties": {
                    "code": { "type": "string" },
                    "messageKey": { "type": "string" },
                    "commandId": {
                      "anyOf": [{ "type": "string", "format": "uuid" }, { "type": "null" }]
                    },
                    "correlationId": { "type": "string", "format": "uuid" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Closed typed validation, lookup, ambiguous command identity or server failure",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": ["code", "messageKey", "commandId", "correlationId"],
                  "properties": {
                    "code": { "type": "string" },
                    "messageKey": { "type": "string" },
                    "commandId": {
                      "anyOf": [{ "type": "string", "format": "uuid" }, { "type": "null" }]
                    },
                    "correlationId": { "type": "string", "format": "uuid" }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Closed typed validation, lookup, ambiguous command identity or server failure",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": ["code", "messageKey", "commandId", "correlationId"],
                  "properties": {
                    "code": { "type": "string" },
                    "messageKey": { "type": "string" },
                    "commandId": {
                      "anyOf": [{ "type": "string", "format": "uuid" }, { "type": "null" }]
                    },
                    "correlationId": { "type": "string", "format": "uuid" }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Closed typed validation, lookup, ambiguous command identity or server failure",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": ["code", "messageKey", "commandId", "correlationId"],
                  "properties": {
                    "code": { "type": "string" },
                    "messageKey": { "type": "string" },
                    "commandId": {
                      "anyOf": [{ "type": "string", "format": "uuid" }, { "type": "null" }]
                    },
                    "correlationId": { "type": "string", "format": "uuid" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/access/callback": {
      "get": {
        "summary": "OIDC callback: single-use state, browser binding, PKCE, nonce, issuer, audience and signature validation",
        "parameters": [
          { "in": "query", "name": "state", "required": false, "schema": { "type": "string" } },
          { "in": "query", "name": "code", "required": false, "schema": { "type": "string" } },
          { "in": "query", "name": "iss", "required": false, "schema": { "type": "string" } },
          { "in": "query", "name": "error", "required": false, "schema": { "type": "string" } },
          {
            "in": "query",
            "name": "session_state",
            "required": false,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "302": { "description": "Authorized local return route or generic login failure" }
        }
      }
    },
    "/api/v1/access/logout": {
      "post": {
        "summary": "End ERP session only",
        "security": [{ "erpSession": [], "csrfToken": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": [],
                "properties": {}
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Redirect target",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": ["url"],
                  "properties": { "url": { "type": "string" } }
                }
              }
            }
          },
          "401": {
            "description": "Session expired",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": ["code", "messageKey", "commandId", "correlationId"],
                  "properties": {
                    "code": { "type": "string" },
                    "messageKey": { "type": "string" },
                    "commandId": {
                      "anyOf": [{ "type": "string", "format": "uuid" }, { "type": "null" }]
                    },
                    "correlationId": { "type": "string", "format": "uuid" }
                  }
                }
              }
            }
          },
          "403": {
            "description": "CSRF or scope denied",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": ["code", "messageKey", "commandId", "correlationId"],
                  "properties": {
                    "code": { "type": "string" },
                    "messageKey": { "type": "string" },
                    "commandId": {
                      "anyOf": [{ "type": "string", "format": "uuid" }, { "type": "null" }]
                    },
                    "correlationId": { "type": "string", "format": "uuid" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}